Fix random-unpause issue (read the rest of the msg)
How to replicate the problem (prior to this commit):
- Open a file.
- Pause.
- Have another person join the party.
- The new person connecting causes your player to unpause...

That's not supposed to happen.

If simulcast-mpv connects fast enough then it could be before mpv
has loaded the file. Once mpv loads the file: it unpauses.
This unpause will send the `Resume` message and then bam:
the other party members are also unpaused.

To account for this, we just force the client to pause if the
party count is greater than 1 and it's the first unpause (hopefully).